+++ tests/component_handlers/streams_bootstrap/test_helm_wrapper.py
@@ -8,11 +8,7 @@ from pytest_mock import MockerFixture
 
 from kpops.cli.pipeline_config import HelmConfig, PipelineConfig
 from kpops.component_handlers.helm_wrapper.config import HelmCommandConfig
-from kpops.component_handlers.helm_wrapper.helm import (
-    Helm,
-    HelmTemplate,
-    load_helm_manifest,
-)
+from kpops.component_handlers.helm_wrapper.helm import Helm, HelmTemplate
 from kpops.component_handlers.helm_wrapper.utils import get_chart
 from kpops.component_handlers.streams_bootstrap.exception import (
     ReleaseNotFoundException,
@@ -331,7 +327,7 @@ class TestHelmWrapper:
 # Resource: chart/templates/test1.yaml
 """
         with pytest.raises(ValueError):
-            helm_templates = list(load_helm_manifest(stdout))
+            helm_templates = list(Helm.load_helm_manifest(stdout))
             assert len(helm_templates) == 0
 
         stdout = """---
@@ -351,7 +347,7 @@ metadata:
             "metadata": {"labels": {"foo": "bar"}},
         }
 
-        helm_templates = list(load_helm_manifest(stdout))
+        helm_templates = list(Helm.load_helm_manifest(stdout))
         assert len(helm_templates) == 1
         helm_template = helm_templates[0]
         assert isinstance(helm_template, HelmTemplate)
@@ -371,7 +367,7 @@ data:
 # Source: chart/templates/test3b.yaml
 foo: bar
 """
-        helm_templates = list(load_helm_manifest(stdout))
+        helm_templates = list(Helm.load_helm_manifest(stdout))
         assert len(helm_templates) == 2
         assert all(
             isinstance(helm_template, HelmTemplate) for helm_template in helm_templates
@@ -399,7 +395,7 @@ data:
 # Source: chart/templates/test3b.yaml
 foo: bar
 """
-        helm_templates = list(load_helm_manifest(stdout))
+        helm_templates = list(Helm.load_helm_manifest(stdout))
         assert len(helm_templates) == 2
         assert all(
             isinstance(helm_template, HelmTemplate) for helm_template in helm_templates
